+++ src/utils/api/MoviesApi.js
@@ -16,7 +16,8 @@ class Api {
 
   async getInitialCards() {
     return fetch(`${this._url}`)
-      .then((res) => this._checkResult(res));
+      .then((res) => this._checkResult(res))
+      .then((res) => res);
   }
 }
